+++ jscomp/bsb/bsb_package_specs.mli
@@ -29,9 +29,6 @@ val from_map : cwd:string -> Ext_json_types.t Map_string.t -> t
 val get_list_of_output_js : t -> string -> string list
 
 val package_flag_of_package_specs : t -> dirname:string -> string
-(**
-   Sample output: {[ -bs-package-output commonjs:lib/js/jscomp/test]}
-*)
 
 (* used to ensure each dir does exist *)
 val list_dirs_by : t -> (string -> unit) -> unit