+++ jscomp/core/js_packages_state.ml
@@ -34,9 +34,6 @@ let make_runtime () : unit =
   make_runtime := true;
   packages_info := Js_packages_info.runtime_package_specs
 
-let make_runtime_test () : unit =
-  packages_info := Js_packages_info.runtime_test_package_specs
-
 let set_package_map module_name =
   (* set_package_name name ;
      let module_name = Ext_namespace.namespace_of_package_name name  in *)