+++ jscomp/core/js_packages_state.mli
@@ -26,8 +26,6 @@ val set_package_name : string -> unit
 
 val make_runtime : unit -> unit
 
-val make_runtime_test : unit -> unit
-
 val set_package_map : string -> unit
 
 val get_packages_info : unit -> Js_packages_info.t