+++ jscomp/test/acyc/a0_a1.js
@@ -1,7 +1,0 @@
-'use strict';
-
-
-var v = 3;
-
-exports.v = v;
-/* No side effect */